+++ pac4j-core/src/main/java/org/pac4j/core/util/Announcement.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
+import static org.pac4j.core.util.Pac4jConstants.EMPTY_STRING;
+
 /**
  * Announcing next versions.
  *
@@ -22,9 +24,23 @@ public final class Announcement {
 
     private final String message;
 
+    private final boolean mailing;
+
+    public Announcement(final String message) {
+        this(null, message, false);
+    }
+
     public Announcement announce() {
         if (!announced) {
-            LOGGER.warn("\u26A0 In version {}, {} (\u2709 contact `pac4j-dev@googlegroups.com` for discussions)", version, message);
+            var contact = " (\u2709 contact `pac4j-dev@googlegroups.com` for discussions)";
+            if (!mailing) {
+                contact = EMPTY_STRING;
+            }
+            var inversion = EMPTY_STRING;
+            if (version != null) {
+                inversion = "In version " + version + ", ";
+            }
+            LOGGER.warn("\u26A0 {}{}{}", inversion, message, contact);
             announced = true;
         }
         return this;